Add customData to Context.
customData can be anything. It is up to individual plugins to define this, and it can include anything a plugin might need to use later, such as whether or not a message was an emote/action. Most plugins that use this will want to set customData to an Object containing the properties they want.